Create Pig Latin Word Translator
Pig Latin takes the first consonant of a word, moves it to the end of the word, and adds on  "ay". 
If a word begins with a vowel you just add "way" to the end.

This is just a beginner-level project, it's not too fancy but I was rather proud to have worked out and coded this by myself.